+++ example/src/bin/client/admin.rs
@@ -31,14 +31,12 @@ pub struct UserListUsersResponse {
 pub struct User {
     pub id: i64,
     pub username: String,
-    pub password: String,
     pub name: String,
     pub age: i32,
     pub score: f64,
     pub balance: f32,
     pub active: bool,
     pub tags: Vec<String>,
-    pub metadata: Option<String>,
     pub role: Role,
     pub status: Status,
     pub profile: Profile,
@@ -109,14 +107,12 @@ pub struct UserUpdateUserRequest {
 pub struct UserUpdateUser {
     pub id: i64,
     pub username: String,
-    pub password: String,
     pub name: String,
     pub age: i32,
     pub score: f64,
     pub balance: f32,
     pub active: bool,
     pub tags: Vec<String>,
-    pub metadata: Option<String>,
     pub role: Role,
     pub status: Status,
     pub profile: Profile,
@@ -146,14 +142,12 @@ pub struct UserDeleteUserIdRequest {
 pub struct UserDeleteUser {
     pub id: i64,
     pub username: String,
-    pub password: String,
     pub name: String,
     pub age: i32,
     pub score: f64,
     pub balance: f32,
     pub active: bool,
     pub tags: Vec<String>,
-    pub metadata: Option<String>,
     pub role: Role,
     pub status: Status,
     pub profile: Profile,
@@ -418,7 +412,6 @@ impl<'a> AdminUserApi<'a> {
                     .map(|_| User {
                         id: r.read_i64(),
                         username: r.read_str(),
-                        password: r.read_str(),
                         name: r.read_str(),
                         age: r.read_i32(),
                         score: r.read_f64(),
@@ -427,11 +420,6 @@ impl<'a> AdminUserApi<'a> {
                         tags: {
                             let _count = r.read_u32() as usize;
                             (0.._count).map(|_| r.read_str()).collect::<Vec<_>>()
-                        },
-                        metadata: if r.read_u8() == 1 {
-                            Some(r.read_str())
-                        } else {
-                            None
                         },
                         role: match r.read_u8() {
                             0 => Role::Admin,
@@ -558,7 +546,6 @@ impl<'a> AdminUserApi<'a> {
             Some(UserUpdateUser {
                 id: r.read_i64(),
                 username: r.read_str(),
-                password: r.read_str(),
                 name: r.read_str(),
                 age: r.read_i32(),
                 score: r.read_f64(),
@@ -567,11 +554,6 @@ impl<'a> AdminUserApi<'a> {
                 tags: {
                     let _count = r.read_u32() as usize;
                     (0.._count).map(|_| r.read_str()).collect::<Vec<_>>()
-                },
-                metadata: if r.read_u8() == 1 {
-                    Some(r.read_str())
-                } else {
-                    None
                 },
                 role: match r.read_u8() {
                     0 => Role::Admin,
@@ -695,7 +677,6 @@ impl<'a> AdminUserApi<'a> {
             Some(UserDeleteUser {
                 id: r.read_i64(),
                 username: r.read_str(),
-                password: r.read_str(),
                 name: r.read_str(),
                 age: r.read_i32(),
                 score: r.read_f64(),
@@ -704,11 +685,6 @@ impl<'a> AdminUserApi<'a> {
                 tags: {
                     let _count = r.read_u32() as usize;
                     (0.._count).map(|_| r.read_str()).collect::<Vec<_>>()
-                },
-                metadata: if r.read_u8() == 1 {
-                    Some(r.read_str())
-                } else {
-                    None
                 },
                 role: match r.read_u8() {
                     0 => Role::Admin,